linux-bootloader: take note of the `image_device_path` in `PeInMemory`
As soon as we open in exclusive our own image, we will not be able easily to re-open it
because of the locking mechanism.

Therefore, we should use the opportunity to get information we may be interested in
and avoid us re-opening the loaded image protocol.

This is useful to know from where you were loaded for example, e.g. filesystem path.